streamlit/requirements.txt
altair==5.0.1
altair-data-server==0.4.1
altair-viewer==0.4.0
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.11.1
blinker==1.6.2
bokeh==3.2.1
bs4==0.0.1
cachetools==5.3.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.4
colorama==0.4.6
compress-pickle==2.1.0
contourpy==1.0.6
cycler==0.11.0
dacite==1.8.1
decorator==5.1.1
distlib==0.3.6
docopt==0.6.2
exceptiongroup==1.1.2
fastapi==0.88.0
filelock==3.12.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
htmlmin==0.1.12
hydralit==1.0.14
hydralit-components==1.0.10
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.2
mdurl==0.1.2
multimethod==1.9.1
networkx==3.1
numpy==1.23.5
packaging==21.3
pandas==1.5.1
pandas-profiling==3.6.6
patsy==0.5.3
phik==0.12.3
Pillow==9.3.0
pipreqs==0.4.13
platformdirs==3.9.1
plotly==5.11.0
portpicker==1.5.2
prettytable==3.8.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pycodestyle==2.11.0
pydantic==1.10.4
pydantic_core==2.1.2
pydeck==0.8.1b0
pyecharts==2.0.3
pygame==2.1.2
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.29.1
requests==2.28.1
rich==13.4.2
rpds-py==0.8.10
scikit-learn==1.1.3
scipy==1.11.1
seaborn==0.12.2
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
statsmodels==0.14.0
streamlit==1.24.1
streamlit-echarts==0.4.0
streamlit-option-menu==0.3.6
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typeguard==2.13.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.13
uvicorn==0.20.0
validators==0.20.0
virtualenv==20.24.0
visions==0.7.5
watchdog==3.0.0
wcwidth==0.2.6
wordcloud==1.9.2
xyzservices==2023.7.0
yarg==0.1.9
ydata-profiling==4.4.0
zipp==3.16.1
